requirements.txt
langgraph>=0.4.8
langgraph-checkpoint>=2.0.26
langgraph-prebuilt>=0.2.2
langgraph-sdk>=0.1.70
langgraph-supervisor>=0.0.1
langgraph-swarm>=0.0.1
langchain-mcp-adapters>=0.0.1
langmem>=0.0.1
agentevals>=0.0.1
langchain>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
ormsgpack>=1.10.0
xxhash>=3.5.0
qdrant-client>=1.7.0
sentence-transformers>=2.5.1
dnspython>=2.4.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
google-auth>=2.16.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.70.0
aiohttp>=3.8.0
msal>=1.20.0
msgraph-core>=0.2.2
clamd>=1.0.2
python-Levenshtein>=0.20.9 